(feature) Add timer to homepage.
This commit adds timer functionality to the homepage, allowing users to
visit the homepage and get a random timer between one and six minutes.

There’s a start button which allows the user to start the timer, a pause
button which allows the user to pause the timer, and a roll button which
allows the user to change the duration of the timer to another random
time between one and six minutes. At the end of the timer an alarm sound
is played.

This feature will help users accomplish their tasks by breaking them up
into manageable micro chunks of time.

This feature is central to this web app, and this is the simplest
possible implementation of it. A much refined user interface, along with
many additional features will come in the future.