(feature) Add ability for users to customize the range of the timer
This commit adds a form to the homepage which allows users to customize
the range of the timer.

The form allows users to input a minimum and maximum, both of which have
to be a whole number between  1 and 25. After entering these numbers the
timer will only roll a duration between those numbers, inclusive on both
ends.

This feature will allow users to better tailor the timer to their
specific tasks and preferences.